Package manual

Package manual

Index

Types

Functions

Simulation

Inference

Utilities

Visualisation

Custom

References

Index

DPOMPs.DPOMPModel

DPOMPs.SimResults

DPOMPs.gillespie_sim

DPOMPs.plot_parameter_heatmap

DPOMPs.plot_parameter_marginal

DPOMPs.plot_parameter_trace

DPOMPs.plot_trajectory

DPOMPs.run_arq_mcmc_analysis

DPOMPs.run_mbp_ibis_analysis

DPOMPs.run_mcmc_analysis

DPOMPs.run_smc2_analysis

Types

DPOMPs.DPOMPModel — Type

DPOMPModel

Fields

model_name – string, e,g, "SIR".

initial_condition – initial condition

rate_function – event rate function.

m_transition – transition matrix.

`observation_function – observation function, use this to add 'noise' to simulated observations.

prior_density – prior density function.

observation_model – observation model likelihood function.

t0_index – index of the parameter that represents the initial time. 0 if fixed at 0.0.

A mutable struct which represents a DSSCT model (see Discuit.jl models for further details).

DPOMPs.SimResults — Type

SimResults

Fields

model_name – string, e,g, "SIR".

particle – the 'trajectory' variable, of type Particle.

population – records the final system state.

observations – simulated observations data (an Array of Observation types.)

The results of a simulation.

Functions

Simulation

DPOMPs.gillespie_sim — Function

gillespie_sim(model, parameters; tmax = 100.0, num_obs = 5)

Run a Doob-Gillespie simulation on model. Returns a SimResults type containing the trajectory and observations data, or an array of the same if n_sims > 1.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

parameters – model parameters.

Optional

tmax – maximum time (default: 100.)

n_obs – the number of observations to draw (default: 5.)

n_sims – number of simulations to draw (default: 1.)

Inference

DPOMPs.run_mcmc_analysis — Function

run_mcmc_analysis(model, obs_data, initial_parameters, steps = 50000, adapt_period = 10000, mbp = true, ppp = 0.3)

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

obs_data – Observations data.

Optional

n_chains – number of Markov chains (optional, default: 3.)

initial_parameters – 2d array of initial model parameters. Each column vector correspondes to a single model parameter.

steps – number of iterations.

adapt_period – number of discarded samples.

mbp – model based proposals (MBP). Set mbp = false for standard proposals.

ppp – the proportion of parameter (vs. trajectory) proposals in Gibbs sampler. Default: 30%. NB. not required for MBP.

fin_adapt – finite adaptive algorithm. The default is false, i.e. [fully] adaptive.

Run an n_chains-MCMC analysis. The initial_parameters are sampled from the prior distribution unless otherwise specified by the user.

A Gelman-Rubin convergence diagnostic is automatically carried out for n_chains > 1 and included in the [multi-chain] results.

Otherwise the results of a single-chain analysis are returned, which include the Geweke test statistics computed for that analysis.

DPOMPs.run_mbp_ibis_analysis — Function

run_mbp_ibis_analysis(model, obs_data, initial_parameters, ess_rs_crit = 0.5; n_props = 3, ind_prop = false, alpha = 1.002)

Run an MBP IBIS analysis based on model and obs_data of type Observations.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

obs_data – Observations data.

np – number of particles (default = 2000.)

ess_rs_crit – resampling criteria (default = 0.5.)

n_props – MBP mutations per step (default = 3.)

ind_prop – true for independent theta proposals (default = false.)

alpha – user-defined, increase for lower acceptance rate targeted (default = 1.002.)

DPOMPs.run_smc2_analysis — Function

run_smc2_analysis(model, obs_data, initial_parameters, ess_rs_crit = 0.5; n_props = 3, ind_prop = false, alpha = 1.002)

Run an SMC^2 (i.e. particle filter IBIS) analysis based on model and obs_data of type Observations.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

obs_data – Observations data.

np – number of particles (default = 2000.)

ess_rs_crit – resampling criteria (default = 0.5.)

n_props – MBP mutations per step (default = 3.)

ind_prop – true for independent theta proposals (default = false.)

alpha – user-defined, increase for lower acceptance rate targeted (default = 1.002.)

DPOMPs.run_arq_mcmc_analysis — Function

run_arq_mcmc_analysis(model::ARQModel; ... )

Run ARQMCMC analysis with chains Markov chains, where n_chains > 1 the Gelman-Rubin convergence diagnostic is also run.

Parameters

model – ARQModel (see docs.)

Named parameters

sample_resolution – i.e. the length of each dimension in the importance sample.

sample_limit – sample limit, should be increased when the variance of model.pdf is high (default: 1.)

n_chains – number of Markov chains (default: 3.)

steps – number of iterations.

burnin – number of discarded samples.

tgt_ar – acceptance rate (default: 0.33.)

run_arq_mcmc_analysis(model, ; ... )

Run ARQMCMC analysis with chains Markov chains, where n_chains > 1 the Gelman-Rubin convergence diagnostic is also run.

Parameters

model – DPOMPModel (see docs.)

Named parameters

sample_resolution – i.e. the length of each dimension in the importance sample.

sample_limit – sample limit, should be increased when the variance of model.pdf is high (default: 3.)

n_chains – number of Markov chains (default: 3.)

steps – number of iterations.

burnin – number of discarded samples.

tgt_ar – acceptance rate (default: 0.33.)

np – number of SMC particles in PF (default: 200.)

ess_crit – acceptance rate (default: 0.33.)

Utilities

Visualisation

DPOMPs.plot_trajectory — Function

plot_trajectory(x)

Parameters

x – SimResults, i.e. from a call to gillespie_sim.

Plot the trajectory of a a DGA simulation on model using UnicodePlots.jl.

DPOMPs.plot_parameter_trace — Function

plot_parameter_trace(mcmc, parameter)

Parameters

sample – MCMCSample, ARQMCMCSample or RejectionSample e.g. from a call to ADD XREF.

parameter – the index of the model parameter to be plotted.

Trace plot of samples from n MCMC analyses for a given model parameter using UnicodePlots.jl.

DPOMPs.plot_parameter_marginal — Function

plot_parameter_marginal(sample, parameter)

Plot the marginal distribution of samples from an MCMC analysis for a given model parameter using UnicodePlots.jl.

Parameters

results – Results object, e.g. of type MCMCSample.

parameter – the index of the model parameter to be plotted.

adapt_period– Adaptation period to be discarded, only required for RejectionSample.

Optional

use_is – Resample IS rather than using MCMC [re]samples (ARQMCMCSample results only.)

DPOMPs.plot_parameter_heatmap — Function

plot_parameter_heatmap(mcmc, x_parameter, y_parameter)

Parameters

mcmc – MCMCResults, e.g. from a call to run_met_hastings_mcmc.

x_parameter – the index of the model parameter to be plotted on the x axis.

y_parameter – the index of the model parameter to be plotted on the y axis.

Plot the marginal distribution of samples from an MCMC analysis for two model parameters using UnicodePlots.jl.

Custom

References

TBA