Package manual

Package manual

Index

Types

Model types

Results

Functions

Models

Simulation

Bayesian parameter inference

Bayesian model analysis

Utilities

Visualisation

Advanced features.

Index

DPOMPs.ARQMCMCSample

DPOMPs.DPOMPModel

DPOMPs.Event

DPOMPs.ImportanceSample

DPOMPs.MCMCSample

DPOMPs.Observation

DPOMPs.Particle

DPOMPs.RejectionSample

DPOMPs.SimResults

DPOMPs.generate_custom_model

DPOMPs.generate_custom_particle

DPOMPs.generate_model

DPOMPs.get_observations

DPOMPs.gillespie_sim

DPOMPs.plot_model_evidence

DPOMPs.plot_parameter_heatmap

DPOMPs.plot_parameter_marginal

DPOMPs.plot_parameter_trace

DPOMPs.plot_trajectory

DPOMPs.print_results

DPOMPs.run_arq_mcmc_analysis

DPOMPs.run_custom_mcmc_analysis

DPOMPs.run_mbp_ibis_analysis

DPOMPs.run_mcmc_analysis

DPOMPs.run_model_comparison_analysis

DPOMPs.run_smc2_analysis

DPOMPs.tabulate_results

Types

Model types

DPOMPs.DPOMPModel — Type

DPOMPModel

A mutable struct which represents a DSSCT model (see Models for further details).

Fields

model_name – string, e,g, "SIR".

rate_function – event rate function.

initial_condition – initial condition.

m_transition – transition matrix.

`obs_function – observation function, use this to add 'noise' to simulated observations.

obs_model – observation model likelihood function.

prior – prior [multivariate] Distributions.Distribution.

t0_index – index of the parameter that represents the initial time. 0 if fixed at 0.0.

DPOMPs.Particle — Type

Particle

E.g. the main results of a simulation including the initial and final conditions, but not the full state trajectory.

Fields

theta – e.g. simulation parameters.

initial_condition – initial system state.

final_condition – final system state.

trajectory – the event history.

log_like – trajectory log likelihood, mainly for internal use.

DPOMPs.Event — Type

Event

Requires no explanation.

Fields

time – the time of the event.

event_type – indexes the rate function and transition matrix.

DPOMPs.Observation — Type

Observation

A single observation. Note that by default val has the same size as the model state space. However that is not necessary - it need only be compatible with the observation model.

Fields

time – similar to Event.time, the time of the observation.

obs_id – <1 if not a resampling step.

prop – optional information for the observation model.

val – the observation value.

Results

DPOMPs.SimResults — Type

SimResults

The results of a simulation, including the full state trajectory.

Fields

model_name – string, e,g, "SIR".

particle – the 'trajectory' variable, of type Particle.

population – records the final system state.

observations – simulated observations data (an Array of Observation types.)

DPOMPs.ImportanceSample — Type

ImportanceSample

The results of an importance sampling analysis, such as iterative batch importance sampling algorithms.

Fields

mu – weighted sample mean.

cv – weighted covariance.

theta – two dimensional array of samples, e.g. parameter; iteration.

weight – sample weights.

run_time – application run time.

bme – Estimate (or approximation) of the Bayesian model evidence.

DPOMPs.RejectionSample — Type

RejectionSample

Essentially, the main results of an MCMC analysis, consisting of samples, mean, and covariance matrix.

Fields

samples – three dimensional array of samples, e.g. parameter; iteration; Markov chain.

mu – sample mean.

cv – sample covariance matrix.

DPOMPs.MCMCSample — Type

MCMCSample

The results of an MCMC analysis, mainly consisting of a RejectionSample.

Fields

samples – samples of type RejectionSample.

adapt_period – adaptation (i.e. 'burn in') period.

sre – scale reduction factor estimate, i.e. Gelman diagnostic.

run_time – application run time.

DPOMPs.ARQMCMCSample — Type

ARQMCMCSample

The results of an ARQ MCMC analysis including the ImportanceSample and resampled RejectionSample.

The sre scale factor reduction estimates relate the rejection (re)samples to the underlying importance sample.

Fields

imp_sample – main results, i.e. ImportanceSample.

samples – resamples, of type RejectionSample.

adapt_period – adaptation (i.e. 'burn in') period.

sample_resolution – number of distinct [possible] sample values along each dimension in the unit cube.

sample_limit – maximum number of samples per theta tupple.

grid_range – bounds of the parameter space.

sre – scale reduction factor estimate, i.e. Gelman diagnostic. NB. only valid for resamples.

run_time – application run time.

Functions

Models

DPOMPs.generate_model — Function

generate_model(model_name, initial_condition; freq_dep = false, obs_error = 2.0)

Generates an DPOMPModel instance. Observation models are generated using the partial_gaussian_obs_model function, with `σ = obs_error (see that functions entry for further details.)

Parameters

model_name – the model, e.g. "SI"; "SIR"; "SEIR"; etc

initial_condition – initial condition.

Optional parameters

freq_dep – epidemiological models only, set to true for frequency-dependent contact rates.

obs_error – average observation error (default = 2.)

model_name options

"SI"

"SIR"

"SIS"

"SEI"

"SEIR"

"SEIS"

"SEIRS"

"PREDPREY"

"ROSSMAC"

Examples

generate_model("SIS", [100,1])

DPOMPs.generate_custom_model — Function

generate_custom_model(model_name, rate_function, initial_condition, m_transition; ... )

Generates an DPOMPModel instance. Observation models are generated using the partial_gaussian_obs_model function, with `σ = obs_error (see that functions entry for further details.)

Parameters

model_name – the model, e.g. "SIR", "SEIR-custom", etc.

rate_function – event rate function.

initial_condition – initial condition

m_transition – transition matrix.

Optional parameters

`observation_function – observation function, use this to add 'noise' to simulated observations.

obs_error – average observation error (default = 2.)

observation_model – observation model likelihood function.

prior_density – prior density function.

t0_index – index of the parameter that represents the initial time. 0 if fixed at 0.0.

Examples

generate_custom_model("SIS", [100,1])

Missing docstring.

Missing docstring for partial_gaussian_obs_model. Check Documenter's build log for details.

Simulation

DPOMPs.gillespie_sim — Function

gillespie_sim(model, parameters; tmax = 100.0, num_obs = 5)

Run a Doob-Gillespie simulation on model. Returns a SimResults type containing the trajectory and observations data, or an array of the same if n_sims > 1.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

parameters – model parameters.

Optional

tmax – maximum time (default: 100.)

n_obs – the number of observations to draw (default: 5.)

n_sims – number of simulations to draw (default: 1.)

Example

using DPOMPs
m = generate_model("SIR", [50, 1, 0])
x = DPOMPs.gillespie_sim(model, [0.005, 0.12])
println(DPOMPs.plot_trajectory(x))

Bayesian parameter inference

DPOMPs.run_mcmc_analysis — Function

run_mcmc_analysis(model, obs_data; ... )

Run an n_chains-MCMC analysis. The initial_parameters are sampled from the prior distribution unless otherwise specified by the user.

A Gelman-Rubin convergence diagnostic is automatically carried out for n_chains > 1 and included in the [multi-chain] analysis results.

Otherwise the results of a single-chain analysis are returned, which include the Geweke test statistics computed for that analysis.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

obs_data – Observations data.

Optional

n_chains – number of Markov chains (optional, default: 3.)

initial_parameters – 2d array of initial model parameters. Each column vector correspondes to a single model parameter.

steps – number of iterations.

adapt_period – number of discarded samples.

mbp – model based proposals (MBP). Set mbp = false for standard proposals.

ppp – the proportion of parameter (vs. trajectory) proposals in Gibbs sampler. Default: 30%. NB. not required for MBP.

fin_adapt – finite adaptive algorithm. The default is false, i.e. [fully] adaptive.

mvp – increase for a higher proportion of 'move' proposals. NB. not applicable if MBP = true (default: 2.)

Example

y = x.observations                          # some simulated data
model = generate_model("SIR", [50, 1, 0])   # a model
results = run_mcmc_analysis(model, y; fin_adapt = true) # finite-adaptive MCMC
tabulate_results(results)                   # optionally, show the results

DPOMPs.run_mbp_ibis_analysis — Function

run_mbp_ibis_analysis(model, obs_data; ... )

Run an MBP IBIS analysis based on model and obs_data of type Observations.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

obs_data – Observations data.

Optional

np – number of particles (default = 4000.)

ess_rs_crit – resampling criteria (default = 0.5.)

n_props – MBP mutations per step (default = 3.)

ind_prop – true for independent theta proposals (default = false.)

alpha – user-defined, increase for lower acceptance rate targeted (default = 1.002.)

Example

# NB. using 'y' and 'model' as above
results = run_mbp_ibis_analysis(model, y)# importance sample
tabulate_results(results)                # show the results

DPOMPs.run_smc2_analysis — Function

run_smc2_analysis(model, obs_data; ... )

Run an SMC^2 (i.e. particle filter IBIS) analysis based on model and obs_data of type Observations.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

obs_data – Observations data.

Optional

np – number of [outer, i.e. theta] particles (default = 2000.)

npf – number of [inner] particles (default = 200.)

ess_rs_crit – resampling criteria (default = 0.5.)

ind_prop – true for independent theta proposals (default = false.)

alpha – user-defined, increase for lower acceptance rate targeted (default = 1.002.)

Example

# NB. using 'y' and 'model' as above
results = run_smc2_analysis(model, y)   # importance sample
tabulate_results(results)               # show the results

DPOMPs.run_arq_mcmc_analysis — Function

run_arq_mcmc_analysis(model::ARQModel; ... )

Run ARQMCMC analysis with chains Markov chains, where n_chains > 1 the Gelman-Rubin convergence diagnostic is also run.

Parameters

model – ARQModel (see docs.)

Named parameters

sample_resolution – i.e. the length of each dimension in the importance sample.

sample_limit – sample limit, should be increased when the variance of model.pdf is high (default: 1.)

n_chains – number of Markov chains (default: 3.)

steps – number of iterations.

burnin – number of discarded samples.

tgt_ar – acceptance rate (default: 0.33.)

run_arq_mcmc_analysis(model, obs_data, theta_range; ... )

Run ARQMCMC analysis with chains Markov chains, where n_chains > 1 the Gelman-Rubin convergence diagnostic is also run.

Parameters

model – DPOMPModel (see docs.)

obs_data – Observations data.

theta_range – An array giving the bounds of the parameter space.

Optional

sample_resolution – i.e. the length of each dimension in the importance sample.

sample_limit – sample limit, should be increased when the variance of model.pdf is high (default: 3.)

n_chains – number of Markov chains (default: 3.)

steps – number of iterations.

burnin – number of discarded samples.

tgt_ar – acceptance rate (default: 0.33.)

np – number of SMC particles in PF (default: 200.)

ess_crit – acceptance rate (default: 0.33.)

Bayesian model analysis

DPOMPs.run_model_comparison_analysis — Function

run_model_comparison_analysis(model, obs_data; ... )

Run n_runs independent analyses for each DPOMPModel element in models, and compute [estimate] the Bayesian model evidence (BME.)

Returns an object of type ModelComparisonResults, which includes the mean and standard deviation of the estimates obtained.

Parameters

models – An Array of DPOMPModel.

obs_data – An array of type Observation.

Optional

n_runs – number of algorithm runs used to estimate the bme (and variance.)

algorithm – String representing the inference method used for the analysis, e.g. "SMC2" for SMC^2 (default); "MBPI" for MBP-IBIS; or "ARQ".

Inference algorithm parameters

np – number of [outer, i.e. theta] particles used in IBIS procedures (doesn't apply to ARQ-MCMC.)

ess_rs_crit – Effective sampling size (resampling )

Example

# NB. define some models to compare, e.g. as m1, m2, etc
models = [m1, m2, m3]
results = run_model_comparison_analysis(models, y; n_runs = 10)
tabulate_results(results)   # show the results (optional)

Utilities

DPOMPs.get_observations — Function

get_observations(source)

Parameters

source – DataFrame or filepath (i.e. String) containing the data (with times in the first column).

Create and return an array of type Observation based on a two dimensional array, DataFrame or file location.

DPOMPs.tabulate_results — Function

tabulate_results

Parameters

results – a data structure of type MCMCSample, ImportanceSample or ARQMCMCSample.

Display the results of an inference analysis.

DPOMPs.print_results — Function

print_results

Print the results of a parameter inference analysis to file.

Parameters

samples – a data structure of type SimResults, ImportanceSample, MCMCSample or ARQMCMCSample.

dpath – the directory where the results will be saved.

Visualisation

DPOMPs.plot_trajectory — Function

plot_trajectory(x)

Parameters

x – SimResults, i.e. from a call to gillespie_sim.

Plot the trajectory of a a DGA simulation on model using UnicodePlots.jl.

DPOMPs.plot_parameter_trace — Function

plot_parameter_trace(mcmc, parameter)

Parameters

sample – MCMCSample, ARQMCMCSample or RejectionSample e.g. from a call to ADD XREF.

parameter – the index of the model parameter to be plotted.

Trace plot of samples from n MCMC analyses for a given model parameter using UnicodePlots.jl.

DPOMPs.plot_parameter_marginal — Function

plot_parameter_marginal(sample, parameter)

Plot the marginal distribution of samples from an MCMC analysis for a given model parameter using UnicodePlots.jl.

Parameters

results – Results object, e.g. of type MCMCSample.

parameter – the index of the model parameter to be plotted.

adapt_period– Adaptation period to be discarded, only required for RejectionSample.

Optional

use_is – Resample IS rather than using MCMC [re]samples (ARQMCMCSample results only.)

DPOMPs.plot_parameter_heatmap — Function

plot_parameter_heatmap(mcmc, x_parameter, y_parameter)

Parameters

mcmc – MCMCResults, e.g. from a call to run_met_hastings_mcmc.

x_parameter – the index of the model parameter to be plotted on the x axis.

y_parameter – the index of the model parameter to be plotted on the y axis.

Plot the marginal distribution of samples from an MCMC analysis for two model parameters using UnicodePlots.jl.

DPOMPs.plot_model_evidence — Function

plot_model_evidence(results; boxplot = true)

Plot the Bayesian model evidence (BME) from a model comparison analysis, using UnicodePlots.jl.

Parameters

results – ModelComparisonResults, i.e. from a call to run_model_comparison_analysis.

boxplot – true for a series of boxplots, else a simple UnicodePlots.barplot showing only the average BME for each model variant (default.)

Advanced features.

This section covers functionality for customising the algorithms themselves.

DPOMPs.run_custom_mcmc_analysis — Function

run_custom_mcmc_analysis(model, obs_data, trajectory_prop, [x0_prop], ... )

Run an n_chains data-augmented MCMC analysis, based on the Gibbs sampler with a user defined proposal function.

A function for conditionally sampling the initial trajectory variable can optionally be specified, use the Doob-Gillespie algorithm is used by default.

Elsewise, this function is equivalent to calling runmcmcanalysis with mbp = false, which invokes the standard Gibbs sampler.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

obs_data – Observations data.

trajectory_prop – .

x0_prop – Initial state variable sampler, DGA by default.

Optional

n_chains – number of Markov chains (optional, default: 3.)

initial_parameters – 2d array of initial model parameters. Each column vector correspondes to a single model parameter.

steps – number of iterations.

adapt_period – number of discarded samples.

ppp – the proportion of parameter (vs. trajectory) proposals in Gibbs sampler. Default: 0.3, or 30%.

fin_adapt – finite adaptive algorithm. The default is false, i.e. [fully] adaptive.

DPOMPs.generate_custom_particle — Function

generate_custom_particle(model, obs_data, trajectory_prop, [x0_prop], ... )

For use with run_custom_mcmc_analysis(). Initialises a Particle based on an array of type Event. Also evaluates the likelihood function.

Parameters

model – DPOMPModel (see [DCTMPs.jl models]@ref).

obs_data – Observations data.

trajectory – Array of Event types.

Optional

theta – model parameters, sampled from prior unless otherwise specified.